move interpolation function used from proc_set_xdt out of _msv2, #461
