+++ tests/unit/measurement_set/_utils/_msv2/test_msv4_sub_xdss.py
@@ -55,27 +55,6 @@ def test_rename_and_interpoalte_to_time_main(msv4_xds_min):
     xr.testing.assert_equal(out_xds.time, input_time)
 
 
-def test_interpolate_to_time_bogus(antenna_xds_min, msv4_xds_min):
-    from xradio.measurement_set._utils._msv2.msv4_sub_xdss import interpolate_to_time
-
-    input_time = msv4_xds_min.time
-    out_xds = interpolate_to_time(
-        antenna_xds_min, interp_time=input_time, message_prefix="test_call"
-    )
-    assert out_xds == input_time
-
-
-def test_interpolate_to_time_main(msv4_xds_min):
-    from xradio.measurement_set._utils._msv2.msv4_sub_xdss import interpolate_to_time
-
-    input_time = msv4_xds_min.time
-    out_xds = interpolate_to_time(
-        msv4_xds_min, interp_time=input_time, message_prefix="test_call"
-    )
-
-    xr.testing.assert_equal(out_xds.time, input_time)
-
-
 @pytest.fixture(scope="session")
 def ant_xds_station_name_ids():
     """Makes a ant_xds_station_name_ids as the one used in conversion (it still has antenna_id coord)"""
